Fix libavresample detection
Since the build scripts were rewritten with CMake, libavresample has not
been used by libavg because the necessary defines were never set. Fix
this by adding a new define to avgconfig.h and using it.

This is required for FFmpeg 4.0 support because upstream FFmpeg has
removed the old ReSampleContext API.